% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{combine}
\alias{combine}
\title{the function that groups together the quantitative and qualitative variables}
\usage{
combine(qual, quant)
}
\arguments{
\item{qual}{qualitative variables}

\item{quant}{quantitative variables}
}
\value{
a matrix of predictor variables
}
\description{
the function that groups together the quantitative and qualitative variables
}
