% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{controlerBatch}
\alias{controlerBatch}
\title{Batch_Size's Control function}
\usage{
controlerBatch(batch_size, donnees, mode)
}
\arguments{
\item{batch_size}{is the number of observations for the mini-batch mode}

\item{donnees}{is the dataset}

\item{mode}{the mode chosen to be used to update coefficients of stochastic gradient descent}
}
\value{
Error message:"The batch-size must be positive and does not exceed the number of data"
}
\description{
Batch_Size's Control function
}
