% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{controlerDataType}
\alias{controlerDataType}
\title{DataType's Control function}
\usage{
controlerDataType(data)
}
\arguments{
\item{data}{the dataframe containing the variables in the model}
}
\value{
Error message:"The input is not a dataframe !"
}
\description{
DataType's Control function
}
