% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{controlerFormula}
\alias{controlerFormula}
\title{Formula's Control function}
\usage{
controlerFormula(formula)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted}
}
\value{
Error message:"Formula must be R formula !"
}
\description{
Formula's Control function
}
