% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{controlerItermax}
\alias{controlerItermax}
\title{max_iter's Control function}
\usage{
controlerItermax(max_iter)
}
\arguments{
\item{max_iter}{is the number of iterations}
}
\value{
Error message:"max_iter must imperatively be positive"
}
\description{
max_iter's Control function
}
