% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{controlerMode}
\alias{controlerMode}
\title{Mode's Control function}
\usage{
controlerMode(mode)
}
\arguments{
\item{mode}{the mode chosen to be used to update coefficients of stochastic gradient descent}
}
\value{
Error message:"Incorrect mode" and the list of possible modes
}
\description{
Mode's Control function
}
