% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{controlerNombreModalites}
\alias{controlerNombreModalites}
\title{the target modality control function}
\usage{
controlerNombreModalites(y)
}
\arguments{
\item{y}{the target variable}
}
\value{
Error message:"The target variable has more than 2 modalities"
}
\description{
the target modality control function
}
