% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticReg_GradDesc.R
\name{EvolEta}
\alias{EvolEta}
\title{the function of changes in the learning rate}
\usage{
EvolEta(learning_rate, nb_iter)
}
\arguments{
\item{learning_rate}{learning rate for gradient descent}

\item{nb_iter}{number of iterationn}
}
\value{

}
\description{
the function of changes in the learning rate
}
