% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticReg_GradDesc.R
\name{FonctionCout}
\alias{FonctionCout}
\title{Logistic regression loss function}
\usage{
FonctionCout(expli, cible, coeff)
}
\arguments{
\item{expli}{the predictor variables}

\item{cible}{the target (the variable to predict)}

\item{coeff}{the logistic regression coefficients to be estimated}
}
\value{
a vector
}
\description{
Logistic regression loss function
}
