% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{formule_donnees_exp}
\alias{formule_donnees_exp}
\title{Function to extract predictor variables from formula}
\usage{
formule_donnees_exp(formula, donnees)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted}

\item{donnees}{the dataset}
}
\value{
the explanatory variables
}
\description{
Function to extract predictor variables from formula
}
