% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{formule_extract_cible}
\alias{formule_extract_cible}
\title{Function to extract the target from formula}
\usage{
formule_extract_cible(formula, donnees)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted}

\item{donnees}{the dataset}
}
\value{
the target variable
}
\description{
Function to extract the target from formula
}
