% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{getmode}
\alias{getmode}
\title{Function that replaces the NA with the mode}
\usage{
getmode(v)
}
\arguments{
\item{v}{a vector containing the NA values}
}
\value{

}
\description{
Function that replaces the NA with the mode
}
