% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticReg_GradDesc.r
\name{GradDescente}
\alias{GradDescente}
\title{Logistic regression function with gradient descent}
\usage{
GradDescente(
  x,
  y,
  eta = 0.3,
  max_iter = 500,
  tol = 0.001,
  mode_desc = "batch",
  batch_size = 32,
  nc
)
}
\arguments{
\item{x}{the predictor variables}

\item{y}{the target (the variable to predict)}

\item{eta}{the learning rate for gradient descent. Default value: 0.3}

\item{max_iter}{is the number of iterations,default value:: 500}

\item{tol}{is the minimum movement allowed for each iteration.Default value: 1e-3}

\item{mode_desc}{the mode chosen to be used to update coefficients of stochastic gradient descent.We have 3 posiblities: {« batch », « online », « mini_batch »}}

\item{batch_size}{the number of observations for the mini-batch mode; Default value: 32}

\item{nc}{indicates the number of cores to be used in parallel programming}
}
\value{
a list of (theta(coefficients), a  number of iteration and deviances)
}
\description{
Logistic regression function with gradient descent
}
