% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticReg_GradDesc.r
\name{Gradient}
\alias{Gradient}
\title{Gradient function}
\usage{
Gradient(expli, cible, coeff)
}
\arguments{
\item{expli}{the predictor variables}

\item{cible}{the target (the variable to predict)}

\item{coeff}{the logistic regression coefficients to be estimated}
}
\value{
a gradient vector
}
\description{
Gradient function
}
