% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticReg_GradDesc.R
\name{MajCoefficient}
\alias{MajCoefficient}
\title{Update coefficients function}
\usage{
MajCoefficient(expli, cibles, coeff, learning_rate, tolerance, convergence)
}
\arguments{
\item{expli}{the predictor variables}

\item{cibles}{the target (the variable to predict)}

\item{coeff}{the logistic regression coefficients to be estimated}

\item{learning_rate}{is the learning rate for gradient descent}

\item{tolerance}{is the minimum movement allowed for each iteration}

\item{convergence}{logical true/false}
}
\value{
a list containing updated coefficients and convergence status
}
\description{
Update coefficients function
}
