% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{ncores}
\alias{ncores}
\title{Function to control and get cores}
\usage{
ncores(cores_choice)
}
\arguments{
\item{cores_choice}{the number of CPU cores chosen by the user}
}
\value{
cores_choice the number of CPU cores that will be used in parallelisation
}
\description{
Function to control and get cores
}
