% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogRegGD.R
\name{predict}
\alias{predict}
\title{Predict function}
\usage{
predict(objet_Reg, newdata, type)
}
\arguments{
\item{objet_Reg}{is a S3 object provided by the function fit()}

\item{newdata}{the dataframe (DataTest) that we will use to make the prediction}

\item{type}{indicates the type of prediction {“class”: predicted class , “posterior”:belonging class probability}}
}
\value{
Function that return a list (the predicted probs,confusion matrix and error rate) or the belonging class probability
}
\description{
Predict function
}
\examples{
 \dontrun{
 predict(objet,newdata,type= "class")
 predict(objet,newdata,type= "posterior")
 }
}
