% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogRegGD.R
\name{print.ObjectLogRegGD}
\alias{print.ObjectLogRegGD}
\title{the generic "print" method}
\usage{
\method{print}{ObjectLogRegGD}(objet)
}
\arguments{
\item{objet}{is the fitted object}
}
\value{

}
\description{
the generic "print" method
}
\examples{
\dontrun{
 print(objet)
 }
}
