% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticReg_GradDesc.r
\name{RegLogFonction}
\alias{RegLogFonction}
\title{Segmoid funcion}
\usage{
RegLogFonction(a)
}
\arguments{
\item{a}{linear combination of predictor variables}
}
\value{
a vector of probabilities of belonging to the positive or negative class
}
\description{
Segmoid funcion
}
