% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogRegGD.R
\name{summary.ObjectLogRegGD}
\alias{summary.ObjectLogRegGD}
\title{the generic "Summary" method}
\usage{
\method{summary}{ObjectLogRegGD}(objet)
}
\arguments{
\item{objet}{is the fitted object}
}
\value{
summary of the results
}
\description{
the generic "Summary" method
}
\examples{
 \dontrun{
 summary(objet)
 }

}
