% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{tranformDataset}
\alias{tranformDataset}
\title{Checkdata function}
\usage{
tranformDataset(expli, cible)
}
\arguments{
\item{expli}{predictor variables}

\item{cible}{target variable}
}
\value{
x and y tranformed
}
\description{
Checkdata function
}
