% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{transformQuali}
\alias{transformQuali}
\title{Transformation function on qualitative variables}
\usage{
transformQuali(d)
}
\arguments{
\item{d}{dataset containing qualitative predictor variables}
}
\value{
qualitative variables recoded in 0/1
}
\description{
Transformation function on qualitative variables
}
