% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_Data.R
\name{transformQuanti}
\alias{transformQuanti}
\title{Transformation function on quantitative variables}
\usage{
transformQuanti(d)
}
\arguments{
\item{d}{dataset containing qualitative predictor variable}
}
\value{
the NA values of the quantitative variables replaced by the average of each variable
}
\description{
Transformation function on quantitative variables
}
