% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogRegGD.R
\name{fit}
\alias{fit}
\title{Title}
\usage{
fit(formula, data, mode, batch_size = "32", ncores, itermax)
}
\arguments{
\item{formula}{le formula}

\item{data}{les données}

\item{mode}{le mode d'execution}

\item{batch_size}{size of batch}

\item{ncores}{number of CPU cores}

\item{itermax}{iteration max}
}
\value{
test
}
\description{
Title
}
